+++ bin/gulp-dynamic-routing.js
